separation module
